Improving JBSEAM-4375
When we flush the nested conversation in order to be able to switch to the parent conversation,
we passivate Entities that were stored in the nested conversation.

As you can see in ManagedEntityWrapper, when this switch happens, at the end of process, we return
to the nested conversation, but we never activate the Entities again, leaving null all over the place.

This could trigger hard to debug problems (for example, an unused injection was removed and then the whole page
stopped working due to NPEs)